Refactored ZomicASTCompiler to support Javascript
I extracted `ZomicCompilerState.java`, which has no ANTLR runtime dependencies.
I'll have to rewrite ZomicASTCompiler in JS, implementing the listener interface there.

All Zomic unit tests are passing.
